Pesky newlines
It's interesting, looping through the labels.txt file gives us a $line that also has a \n at the end of it, thus a search for "House Finch" becomes just "House". But only on certain species? It seems that 95% of other species weren't affected by this newline. Anyways this fixes it :)