Make ffmpeg record the first audio stream for RTSP
Small fix for now to handle of RTSP streams that may contain more than 1 audio stream within them,  we just tell ffmpeg to use the first audio stream for the recording.

A future enhancement to this is to allow the user to to set what audio stream to just in case there is a scenario where the first one doesn't work for some reason.