In certain situations trying to recover from stalled analysis we can end up with low / no disk space and be unable to "escape." In the event of a backlog, the loop in birdnet_analysis.sh runs continuously, which means analyzed files do not get moved, which means in turn extract_new_birdsounds.sh cannot create mp3 extractions and move the processed files again, which would allow them to be cleaned up by the cron jobs.
This patch is the obvious fix, i.e. to just limit the maximum amount of files
birdnet_analysis.sh can process to 20. This should be a good trade off given
the small overhead of listing files etc before the main loop begins - note that
for the "normal" case of no backlog, birdnet_analysis usually exits after a
single analysis anyway.